feat(events): cleanup, editing, deleting (#107)
* fix: simplify timezone, break apart 1st/last ordinal

* feat: new event page

* creation of events and basic editing

* format

* fix: background on select component

* feat: add edit icon

no editing yet

* feat: base delete event modal

* feat: base event deletion

* feat: add calendar headers

* fix: use correct club validation logic

* feat: base event draft functionality

* feat: make add show correct date, move freq to rrule

* style: fmt

* feat: derive inputFrequency, dayOfTheMonth

* fix: use str for input frequency

* feat: upTo, repeatType

* feat: create at the right time

* fix: use correct formatting for upTo

* refactor: cleanup

---------
