+++ package.json
@@ -26,9 +26,10 @@
 		"@faker-js/faker": "^8.3.1",
 		"@fontsource-variable/source-code-pro": "^5.0.17",
 		"@fontsource/work-sans": "^5.0.16",
+		"@iconify-json/bx": "^1.1.10",
 		"@leodog896/bodyguard": "^1.3.2",
 		"@playwright/test": "^1.41.0",
-		"@prisma/client": "^5.8.1",
+		"@prisma/client": "^5.10.2",
 		"@svelte-put/qr": "^1.0.3",
 		"@sveltejs/adapter-auto": "^3.1.0",
 		"@sveltejs/adapter-node": "^4.0.0",
@@ -60,7 +61,7 @@
 		"postcss-html": "^1.6.0",
 		"prettier": "^3.2.4",
 		"prettier-plugin-svelte": "^3.1.2",
-		"prisma": "^5.8.1",
+		"prisma": "^5.10.2",
 		"rrule": "^2.8.1",
 		"sanitize-html": "^2.11.0",
 		"sass": "^1.70.0",
@@ -77,6 +78,7 @@
 		"type-fest": "^4.9.0",
 		"typescript": "^5.3.3",
 		"ua-parser-js": "^1.0.37",
+		"unplugin-icons": "^0.18.5",
 		"vite": "^5.0.12",
 		"vitest": "^1.2.1",
 		"zod": "^3.22.4",
