+++ src/lib/components/Select.svelte
@@ -4,6 +4,8 @@
 	export let value: string;
 	export let id: string | null = null;
 	export let style = '';
+
+	export let bg = 'var(--bg)';
 </script>
 
 <div {style} class="wrap">
@@ -11,7 +13,7 @@
 		<span class="label">
 			{label}
 		</span>
-		<span class="labelBG">
+		<span style="--background: {bg}" class="labelBG">
 			{label}
 		</span>
 		<select {id} {name} {style} placeholder={label} bind:value on:click on:blur on:focus on:change>
