+++ src/routes/(app)/org/[id=integer]/club/[clubId]/events/add/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Now what are you doing here.</h1>
